POLIO_2071: reorganize campaign API files
