+++ plugins/polio/api/campaigns/views/campaigns.py
@@ -7,809 +7,50 @@ from django.conf import settings
 from django.core.mail import send_mail
 from django.db.models import Max, Min, Prefetch, Q
 from django.db.models.query import QuerySet
-from django.db.transaction import atomic
 from django.http import HttpResponse, JsonResponse, StreamingHttpResponse
 from django.shortcuts import get_object_or_404
-from django.utils import timezone
 from django.utils.timezone import now
-from django.utils.translation import gettext as _
 from django_filters.rest_framework import DjangoFilterBackend  # type: ignore
-from gspread.exceptions import APIError  # type: ignore
 from rest_framework import filters, permissions, serializers, status
 from rest_framework.decorators import action
-from rest_framework.fields import Field
 from rest_framework.request import Request
 from rest_framework.response import Response
-from rest_framework.validators import UniqueValidator
 
 from hat.api.export_utils import Echo, iter_items
 from iaso.api.common import (
     CONTENT_TYPE_CSV,
     CONTENT_TYPE_XLSX,
-    Custom403Exception,
     CustomFilterBackend,
     DeletionFilterBackend,
     ModelViewSet,
 )
-from iaso.models import Group, OrgUnit
-from plugins.polio.api.campaigns.campaign_groups import CampaignNameSerializer
+from iaso.models import OrgUnit
 from plugins.polio.api.campaigns.campaigns_log import (
     log_campaign_modification,
     serialize_campaign,
 )
-from plugins.polio.api.campaigns.vaccine_authorization_missing_email import (
-    missing_vaccine_authorization_for_campaign_email_alert,
-)
+from plugins.polio.api.campaigns.serializers.anonymous import AnonymousCampaignSerializer
+from plugins.polio.api.campaigns.serializers.calendar import CalendarCampaignSerializer
+from plugins.polio.api.campaigns.serializers.campaigns import CampaignSerializer
+from plugins.polio.api.campaigns.serializers.list import ListCampaignSerializer
+from plugins.polio.api.campaigns.serializers.retrieve import RetrieveCampaignSerializer
+from plugins.polio.api.campaigns.serializers.shared import CampaignTypeSerializer
 from plugins.polio.api.lqas_im.lqas_im_dropdowns import CampaignDropDownSerializer
-from plugins.polio.api.rounds.round import RoundScopeSerializer, RoundSerializer
-from plugins.polio.api.shared_serializers import GroupSerializer, OrgUnitSerializer
+from plugins.polio.api.preparedness.serializers import (
+    CampaignPreparednessSpreadsheetSerializer,
+    PreparednessPreviewSerializer,
+)
+from plugins.polio.api.preparedness.utils import get_current_preparedness
 from plugins.polio.export_utils import generate_xlsx_campaigns_calendar, xlsx_file_name
 from plugins.polio.models import (
     Campaign,
-    CampaignGroup,
-    CampaignScope,
     CampaignType,
     CountryUsersGroup,
     Round,
-    RoundScope,
-    SpreadSheetImport,
-    VaccineAuthorization,
 )
-from plugins.polio.models.base import SubActivity, SubActivityScope
-from plugins.polio.preparedness.calculator import get_preparedness_score
-from plugins.polio.preparedness.parser import InvalidFormatError, get_preparedness
 from plugins.polio.preparedness.spreadsheet_manager import (
     Campaign,
-    generate_spreadsheet_for_campaign,
 )
-from plugins.polio.preparedness.summary import preparedness_summary
-from plugins.polio.services.campaign import (
-    delete_old_scopes_after_scope_level_switch,
-    remove_out_of_scope_org_units_from_sub_activities,
-)
-
-
-# Don't display the url for Anonymous users
-class RoundAnonymousSerializer(RoundSerializer):
-    class Meta:
-        model = Round
-        exclude = ["preparedness_spreadsheet_url"]
-
-
-class CampaignScopeSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = CampaignScope
-        fields = ["group", "vaccine"]
-
-    group = GroupSerializer()
-
-
-class CurrentAccountDefault:
-    """
-    May be applied as a `default=...` value on a serializer field.
-    Returns the current user's account.
-    """
-
-    requires_context = True
-
-    def __call__(self, serializer_field):
-        return serializer_field.context["request"].user.iaso_profile.account_id
-
-
-def check_total_doses_requested(vaccine_authorization, nOPV2_rounds, current_campaign):
-    """
-    Check if the total doses requested per round in a campaign is not superior to the allowed doses in the vaccine authorization.
-    It also emails the nopv2 vaccine team about it.
-    """
-    if vaccine_authorization and vaccine_authorization.quantity is not None:
-        campaigns = Campaign.objects.filter(country=vaccine_authorization.country, deleted_at__isnull=True).exclude(
-            pk=current_campaign.pk
-        )
-        total_doses_requested_for_campaigns = 0
-        campaigns_rounds = [c_round for c_round in Round.objects.filter(campaign__in=campaigns)]
-
-        existing_nopv2_rounds = []
-        for r in campaigns_rounds:
-            if "nOPV2" in r.vaccine_names:
-                existing_nopv2_rounds.append(r)
-
-        for r in existing_nopv2_rounds:
-            if r.started_at and r.doses_requested:
-                if vaccine_authorization.start_date <= r.started_at <= vaccine_authorization.expiration_date:
-                    total_doses_requested_for_campaigns += r.doses_requested
-
-        total_doses_requested = 0
-        for c_round in nOPV2_rounds:
-            if c_round.doses_requested is not None:
-                if c_round.started_at >= vaccine_authorization.start_date <= vaccine_authorization.expiration_date:
-                    total_doses_requested += c_round.doses_requested
-
-        if total_doses_requested + total_doses_requested_for_campaigns > vaccine_authorization.quantity:
-            message = f"The total of doses requested {total_doses_requested} is superior to the autorized doses for this campaign {vaccine_authorization.quantity}."
-            if total_doses_requested_for_campaigns > 0:
-                message = f"The total of doses requested {total_doses_requested} and the aggregation of all previous requested doses {total_doses_requested_for_campaigns} are superior to the autorized doses for this campaign {vaccine_authorization.quantity}."
-            raise Custom403Exception(message)
-
-
-class CampaignTypeSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = CampaignType
-        fields = ["id", "name", "slug"]
-
-
-class CampaignSerializer(serializers.ModelSerializer):
-    rounds = RoundSerializer(many=True, required=False)
-    org_unit: Field = OrgUnitSerializer(source="initial_org_unit", read_only=True)
-    top_level_org_unit_name: Field = serializers.SlugRelatedField(source="country", slug_field="name", read_only=True)
-    top_level_org_unit_id: Field = serializers.SlugRelatedField(source="country", slug_field="id", read_only=True)
-    general_status = serializers.SerializerMethodField()
-    grouped_campaigns = serializers.PrimaryKeyRelatedField(
-        many=True, queryset=CampaignGroup.objects.all(), required=False
-    )
-    # Account is filled per default the one of the connected user that update it
-    account: Field = serializers.PrimaryKeyRelatedField(default=CurrentAccountDefault(), read_only=True)
-    has_data_in_budget_tool = serializers.SerializerMethodField(read_only=True)
-    campaign_types = serializers.PrimaryKeyRelatedField(many=True, queryset=CampaignType.objects.all(), required=False)
-    # Vaccines with real scope
-    vaccines = serializers.SerializerMethodField(read_only=True)
-    single_vaccines = serializers.SerializerMethodField(read_only=True)
-    # Integrated campaigns
-    integrated_to = serializers.PrimaryKeyRelatedField(
-        required=False, allow_null=True, read_only=False, queryset=Campaign.objects.all()
-    )
-    integrated_campaigns = serializers.PrimaryKeyRelatedField(
-        many=True, allow_empty=True, required=False, read_only=False, queryset=Campaign.objects.all()
-    )
-
-    scopes = CampaignScopeSerializer(many=True, required=False)
-    obr_name = serializers.CharField(validators=[UniqueValidator(queryset=Campaign.objects.all())])
-
-    def get_vaccines(self, obj):
-        if obj.vaccines_extended_list:
-            return ",".join(obj.vaccines_extended_list)
-        return ""
-
-    def get_single_vaccines(self, obj):
-        return obj.single_vaccines_extended
-
-    def get_top_level_org_unit_name(self, campaign):
-        if campaign.country:
-            return campaign.country.name
-        return ""
-
-    def get_top_level_org_unit_id(self, campaign):
-        if campaign.country:
-            return campaign.country.id
-        return ""
-
-    def get_general_status(self, campaign):
-        now_utc = timezone.now().date()
-        ordered_rounds = list(campaign.rounds.all())
-        ordered_rounds.sort(key=lambda x: x.number, reverse=True)
-        for round in ordered_rounds:
-            if round.ended_at and now_utc > round.ended_at:
-                return _("Round {} ended").format(round.number)
-            if round.started_at and now_utc >= round.started_at:
-                return _("Round {} started").format(round.number)
-        return _("Preparing")
-
-    def get_has_data_in_budget_tool(self, campaign):
-        return campaign.budget_steps.count() > 0
-
-    def validate(self, attrs):
-        data = super().validate(attrs)
-        polio_type = CampaignType.objects.get(name=CampaignType.POLIO)
-        campaign_types = data["campaign_types"] = data.get("campaign_types", [polio_type])
-        integrated_to = data.get("integrated_to", None)
-        is_polio = polio_type in campaign_types
-
-        if is_polio and (integrated_to is not None):
-            raise serializers.ValidationError("Polio campaign cannot be integrated to other campaign ")
-
-        if (not is_polio) and data.get("integrated_campaigns", None):
-            raise serializers.ValidationError("Non Polio campaign cannot have integrated campaigns")
-
-        return data
-
-    def validate_integrated_to(self, integrated_to):
-        if not integrated_to:
-            return integrated_to
-        polio_type = CampaignType.objects.get(name=CampaignType.POLIO)
-        if polio_type not in integrated_to.campaign_types.all():
-            raise serializers.ValidationError(
-                f"Campaign {integrated_to} is not a polio campaign: it cannot have other campaigns integrated to it"
-            )
-        return integrated_to
-
-    def validate_integrated_campaigns(self, value):
-        """Convert None to empty list for many=True field"""
-        if value is None:
-            return []
-        polio_type = CampaignType.objects.get(name=CampaignType.POLIO)
-        integrated_campaigns_with_wrong_type = [c.obr_name for c in value if polio_type in c.campaign_types.all()]
-        if integrated_campaigns_with_wrong_type:
-            raise serializers.ValidationError(
-                f"Found polio campaign(s) in integrated campaigns: {integrated_campaigns_with_wrong_type}"
-            )
-        return value
-
-    @atomic
-    def create(self, validated_data):
-        grouped_campaigns = validated_data.pop("grouped_campaigns", [])
-        rounds = validated_data.pop("rounds", [])
-        campaign_types = validated_data.pop("campaign_types", [])
-        initial_org_unit = validated_data.get("initial_org_unit")
-        obr_name = validated_data["obr_name"]
-        integrated_to = validated_data.pop("integrated_to", None)
-        integrated_campaigns = validated_data.pop("integrated_campaigns", [])
-        account = self.context["request"].user.iaso_profile.account
-
-        campaign_scopes = validated_data.pop("scopes", [])
-        campaign = Campaign.objects.create(
-            # there seems a bug in DRF the account is not in validated data, when not using HiddenField
-            account_id=self.context["request"].user.iaso_profile.account_id,
-            **validated_data,
-        )
-
-        campaign.grouped_campaigns.set(grouped_campaigns)
-
-        # noinspection DuplicatedCode
-        for scope in campaign_scopes:
-            vaccine = scope.get("vaccine", "")
-            org_units = scope.get("group", {}).get("org_units")
-            scope, _ = campaign.scopes.get_or_create(vaccine=vaccine)
-            source_version_id = None
-            name = f"scope {scope.id} for campaign {campaign.obr_name}" + (f" - {vaccine}" if vaccine else "")
-            if org_units:
-                source_version_ids = set([ou.version_id for ou in org_units])
-                if len(source_version_ids) != 1:
-                    raise serializers.ValidationError("All orgunit should be in the same source version")
-                source_version_id = list(source_version_ids)[0]
-            if not scope.group:
-                scope.group = Group.objects.create(name=name, source_version_id=source_version_id)
-            else:
-                scope.group.source_version_id = source_version_id
-                scope.group.name = name
-                scope.group.save()
-
-            scope.group.org_units.set(org_units)
-
-        if campaign_types != campaign.campaign_types:
-            campaign.campaign_types.set(campaign_types)
-
-        campaign.integrated_to = integrated_to
-
-        for round_data in rounds:
-            scopes = round_data.pop("scopes", [])
-            round_serializer = RoundSerializer(data={**round_data, "campaign": campaign.id}, context=self.context)
-            round_serializer.is_valid(raise_exception=True)
-            round = round_serializer.save()
-
-            for scope in scopes:
-                vaccine = scope.get("vaccine", "")
-                org_units = scope.get("group", {}).get("org_units")
-                source_version_id = None
-                if org_units:
-                    source_version_ids = set([ou.version_id for ou in org_units])
-                    if len(source_version_ids) != 1:
-                        raise serializers.ValidationError("All orgunit should be in the same source version")
-                    source_version_id = list(source_version_ids)[0]
-                scope, _ = round.scopes.get_or_create(vaccine=vaccine)
-                name = f"scope {scope.id} for round {round.number} campaign {campaign.obr_name}" + (
-                    f" - {vaccine}" if vaccine else ""
-                )
-                if not scope.group:
-                    scope.group = Group.objects.create(name=name)
-                else:
-                    scope.group.source_version_id = source_version_id
-                    scope.group.name = name
-                    scope.group.save()
-
-                scope.group.org_units.set(org_units)
-        campaign.integrated_campaigns.set(integrated_campaigns)
-        campaign.update_geojson_field()
-        campaign.save()
-        log_campaign_modification(campaign, None, self.context["request"].user)
-
-        # check if the quantity of the vaccines requested is not superior to the authorized vaccine quantity
-        c_rounds = [r for r in campaign.rounds.all()]
-        nOPV2_rounds = []
-        for r in c_rounds:
-            if "nOPV2" in r.vaccine_names:
-                nOPV2_rounds.append(r)
-
-        if initial_org_unit and len(nOPV2_rounds) > 0:
-            try:
-                initial_org_unit = OrgUnit.objects.get(pk=initial_org_unit.pk)
-                vaccine_authorization = VaccineAuthorization.objects.filter(
-                    country=initial_org_unit, status="VALIDATED", account=account, deleted_at__isnull=True
-                )
-                if vaccine_authorization:
-                    check_total_doses_requested(vaccine_authorization[0], nOPV2_rounds, campaign)
-                else:
-                    missing_vaccine_authorization_for_campaign_email_alert(
-                        obr_name, validated_data["initial_org_unit"], account
-                    )
-            except OrgUnit.DoesNotExist:
-                raise Custom403Exception("error:Org unit does not exists.")
-
-        return campaign
-
-    @atomic
-    def update(self, instance: Campaign, validated_data):
-        old_campaign_dump = serialize_campaign(instance)
-        campaign_scopes = validated_data.pop("scopes", [])
-        rounds = validated_data.pop("rounds", [])
-        initial_org_unit = validated_data.get("initial_org_unit")
-        integrated_to = validated_data.pop("integrated_to", None)
-        integrated_campaigns = validated_data.pop("integrated_campaigns", [])
-
-        account = self.context["request"].user.iaso_profile.account
-        separate_scopes_per_round = validated_data.get("separate_scopes_per_round", instance.separate_scopes_per_round)
-        switch_to_scope_per_round = separate_scopes_per_round and not instance.separate_scopes_per_round
-        switch_to_scope_per_campaign = not separate_scopes_per_round and instance.separate_scopes_per_round
-        keep_scope_per_round = separate_scopes_per_round and instance.separate_scopes_per_round
-        keep_scope_per_campaign = not separate_scopes_per_round and not instance.separate_scopes_per_round
-
-        delete_old_scopes_after_scope_level_switch(
-            switch_to_campaign=switch_to_scope_per_campaign,
-            switch_to_round=switch_to_scope_per_round,
-            campaign=instance,
-        )
-
-        if switch_to_scope_per_campaign or keep_scope_per_campaign:
-            for scope in campaign_scopes:
-                vaccine = scope.get("vaccine", "")
-                org_units = scope.get("group", {}).get("org_units")
-                scope, _ = instance.scopes.get_or_create(vaccine=vaccine)
-                source_version_id = None
-                name = f"scope for campaign {instance.obr_name} - {vaccine or ''}"
-                if org_units:
-                    source_version_ids = set([ou.version_id for ou in org_units])
-                    if len(source_version_ids) != 1:
-                        raise serializers.ValidationError("All orgunit should be in the same source version")
-                    source_version_id = list(source_version_ids)[0]
-                if not scope.group:
-                    scope.group = Group.objects.create(name=name, source_version_id=source_version_id)
-                else:
-                    scope.group.source_version_id = source_version_id
-                    scope.group.name = name
-                    scope.group.save()
-
-                scope.group.org_units.set(org_units)
-
-        round_instances = []
-        # find existing round either by id or number
-        # Fixme this is not a partial update because of the dfault
-        for round_data in rounds:
-            round = None
-            if round_data.get("id"):
-                round_id = round_data["id"]
-                round = Round.objects.get(pk=round_id)
-                if round.campaign != instance:
-                    raise serializers.ValidationError({"rounds": "round is attached to a different campaign"})
-            elif round_data.get("number", None) is not None:
-                try:
-                    round = instance.rounds.get(number=round_data.get("number"))
-                except Round.DoesNotExist:
-                    pass
-            # we pop the campaign since we use the set afterward which will also remove the deleted one
-            round_data.pop("campaign", None)
-            scopes = round_data.pop("scopes", [])
-
-            # Replace ReasonForDelay instance with key_name to avoid type error when calling is_valid
-            # Because the serializer is nested, and data is converted at every level of nesting
-            # datelog["reason_for_delay"] is the ReasonForDelay instance, and not the  key_name that was passed by the front-end
-            # So we have to extract the key_name from the instance and re-pass it to the serializer, otherwise we get an error
-            round_datelogs = round_data.pop("datelogs", [])
-            datelogs_with_pk = [
-                {**datelog, "reason_for_delay": datelog["reason_for_delay"].key_name} for datelog in round_datelogs
-            ]
-            round_data["datelogs"] = datelogs_with_pk
-
-            # At this point `round_data` has already been validated and deserialized to complex Django types.
-            # But it will still be passed to `RoundSerializer.data` as if it were native Python datatypes.
-            # So, we convert it back to a native datatype to prevent `RoundSerializer` to raise an error:
-            # "Incorrect type. Expected pk value, received BudgetProcess."…
-            budget_process = round_data.pop("budget_process", None)
-            round_data["budget_process"] = budget_process.pk if budget_process else None
-
-            round_serializer = RoundSerializer(instance=round, data=round_data, context=self.context)
-            round_serializer.is_valid(raise_exception=True)
-            round_instance = round_serializer.save()
-            round_instances.append(round_instance)
-
-            if switch_to_scope_per_round or keep_scope_per_round:
-                for scope in scopes:
-                    vaccine = scope.get("vaccine", "")
-                    org_units = scope.get("group", {}).get("org_units")
-                    source_version_id = None
-                    if org_units:
-                        source_version_ids = set([ou.version_id for ou in org_units])
-                        if len(source_version_ids) != 1:
-                            raise serializers.ValidationError("All orgunit should be in the same source version")
-                        source_version_id = list(source_version_ids)[0]
-                    name = f"scope for round {round_instance.number} campaign {instance.obr_name}" + (
-                        f" - {vaccine}" if vaccine else ""
-                    )
-                    scope, created = round_instance.scopes.get_or_create(vaccine=vaccine)
-                    if not scope.group:
-                        scope.group = Group.objects.create(name=name)
-                    else:
-                        scope.group.source_version_id = source_version_id
-                        scope.group.name = name
-                        scope.group.save()
-
-                    scope.group.org_units.set(org_units)
-
-        submitted_round_ids = set([r.id for r in round_instances])
-        current_round_ids = set(instance.rounds.values_list("id", flat=True))
-
-        # When some rounds need to be deleted, the payload contains only the rounds to keep.
-        # So we have to detect if somebody wants to delete a round to prevent deletion of
-        # rounds linked to budget processes.
-        has_rounds_to_delete = round_instances and len(current_round_ids) > len(submitted_round_ids)
-        if has_rounds_to_delete:
-            rounds_to_delete = instance.rounds.exclude(id__in=submitted_round_ids)
-            if rounds_to_delete.filter(budget_process__isnull=False).exists():
-                raise serializers.ValidationError("Cannot delete a round linked to a budget process.")
-            rounds_to_delete.delete()
-
-        instance.integrated_to = integrated_to
-        instance.rounds.set(round_instances)
-
-        # We have to detect new rounds manually because of the way rounds are associated to the campaign.
-        new_rounds_ids = submitted_round_ids - current_round_ids
-        if new_rounds_ids:
-            for round in instance.rounds.filter(id__in=new_rounds_ids):
-                round.add_chronogram()
-
-        campaign = super().update(instance, validated_data)
-        campaign.integrated_campaigns.set(integrated_campaigns)
-        campaign.update_geojson_field()
-        campaign.save()
-
-        # Check if there are no subactivities with org units not in their current parent round scope
-        remove_out_of_scope_org_units_from_sub_activities(campaign)
-
-        # check if the quantity of the vaccines requested is not superior to the authorized vaccine quantity
-        c_rounds = [r for r in campaign.rounds.all()]
-        nOPV2_rounds = []
-        for r in c_rounds:
-            if "nOPV2" in r.vaccine_names:
-                nOPV2_rounds.append(r)
-        if initial_org_unit and len(nOPV2_rounds) > 0:
-            try:
-                initial_org_unit = OrgUnit.objects.get(pk=initial_org_unit.pk)
-                vaccine_authorization = VaccineAuthorization.objects.filter(
-                    country=initial_org_unit, status="VALIDATED", account=account, deleted_at__isnull=True
-                )
-                if vaccine_authorization:
-                    check_total_doses_requested(vaccine_authorization[0], nOPV2_rounds, campaign)
-            except OrgUnit.DoesNotExist:
-                raise Custom403Exception("error:Org unit does not exists.")
-
-        log_campaign_modification(campaign, old_campaign_dump, self.context["request"].user)
-        return campaign
-
-    class Meta:
-        model = Campaign
-        # TODO in the future specify the fields that need to be returned so we can remove the deprecated fields
-        # fields = "__all__"
-        exclude = ["geojson"]
-
-        read_only_fields = ["creation_email_send_at", "group"]
-
-
-class CampaignTypeIdAndNameSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = CampaignType
-        fields = ["id", "name"]
-
-
-class IntegratedCampaignSerializer(serializers.ModelSerializer):
-    campaign_types = CampaignTypeIdAndNameSerializer(many=True, read_only=True)
-
-    class Meta:
-        model = Campaign
-        fields = ["id", "obr_name", "campaign_types"]
-
-
-class RetrieveCampaignSerializer(CampaignSerializer):
-    # Override Integrated campaigns fields to send obr_name to the front-end
-    integrated_to = CampaignNameSerializer(read_only=True)
-    integrated_campaigns = IntegratedCampaignSerializer(many=True, read_only=True)
-
-    def validate(self, attrs):
-        # Skip parent validation since this serializer doesn't handle write methods
-        return serializers.ModelSerializer.validate(self, attrs)
-
-    def validate_integrated_to(self, value):
-        # Skip parent validation since this serializer doesn't handle write methods
-        return value
-
-    def validate_integrated_campaigns(self, value):
-        # Skip parent validation since this serializer doesn't handle write methods
-        return value
-
-
-class ListCampaignSerializer(CampaignSerializer):
-    """This serializer contains juste enough data for the List view in the web ui"""
-
-    class NestedListRoundSerializer(RoundSerializer):
-        class Meta:
-            model = Round
-            fields = [
-                "id",
-                "number",
-                "started_at",
-                "ended_at",
-            ]
-
-    rounds = NestedListRoundSerializer(many=True, required=False)
-
-    campaign_types = CampaignTypeSerializer(many=True, required=False)
-
-    class Meta:
-        model = Campaign
-        fields = [
-            "id",
-            "epid",
-            "obr_name",
-            "account",
-            "cvdpv2_notified_at",
-            "top_level_org_unit_name",
-            "top_level_org_unit_id",
-            "rounds",
-            "general_status",
-            "grouped_campaigns",
-            "campaign_types",
-            "is_test",
-            "is_preventive",
-            "on_hold",
-            "is_planned",
-        ]
-        read_only_fields = fields
-
-
-class AnonymousCampaignSerializer(CampaignSerializer):
-    rounds = RoundAnonymousSerializer(many=True)
-    campaign_types = CampaignTypeSerializer(many=True, required=False)
-
-    class Meta:
-        model = Campaign
-        fields = [
-            "id",
-            "epid",
-            "obr_name",
-            "gpei_coordinator",
-            "gpei_email",
-            "description",
-            "initial_org_unit",
-            "creation_email_send_at",
-            "onset_at",
-            "cvdpv2_notified_at",
-            "pv_notified_at",
-            "pv2_notified_at",
-            "virus",
-            "scopes",
-            "vaccines",
-            "detection_status",
-            "detection_responsible",
-            "detection_first_draft_submitted_at",
-            "risk_assessment_status",
-            "risk_assessment_responsible",
-            "investigation_at",
-            "risk_assessment_first_draft_submitted_at",
-            "risk_assessment_rrt_oprtt_approval_at",
-            "ag_nopv_group_met_at",
-            "dg_authorized_at",
-            "verification_score",
-            "budget_status",
-            "who_disbursed_to_co_at",
-            "who_disbursed_to_moh_at",
-            "unicef_disbursed_to_co_at",
-            "unicef_disbursed_to_moh_at",
-            "no_regret_fund_amount",
-            "payment_mode",
-            "rounds",
-            "created_at",
-            "updated_at",
-            "district_count",
-            "top_level_org_unit_name",
-            "top_level_org_unit_id",
-            "is_preventive",
-            "account",
-            "outbreak_declaration_date",
-            "campaign_types",
-            "separate_scopes_per_round",
-        ]
-        read_only_fields = fields
-
-
-class CampaignDropDownSerializer(serializers.ModelSerializer):
-    obr_name = serializers.CharField()
-    id = serializers.CharField()
-    campaign_types = CampaignTypeIdAndNameSerializer(many=True, read_only=True)
-
-    class Meta:
-        model = Campaign
-        fields = ["id", "obr_name", "campaign_types"]
-
-
-class CalendarCampaignSerializer(CampaignSerializer):
-    """This serializer contains juste enough data for the Calendar view in the web ui. Read only.
-    Used by both anonymous and non-anonymous user"""
-
-    class NestedSubactivitySerializer(serializers.ModelSerializer):
-        class NestedScopeSerializer(serializers.ModelSerializer):
-            class NestedGroupSerializer(GroupSerializer):
-                class Meta:
-                    model = Group
-                    fields = ["id"]
-
-            class Meta:
-                model = SubActivityScope
-                fields = ["group", "vaccine"]
-
-            group = NestedGroupSerializer()
-
-        scopes = NestedScopeSerializer(many=True)
-        round_number = serializers.IntegerField(source="round.number")
-
-        class Meta:
-            model = SubActivity
-            fields = ["name", "start_date", "end_date", "scopes", "id", "vaccine_names", "round_number"]
-
-    class NestedListRoundSerializer(RoundSerializer):
-        class NestedScopeSerializer(RoundScopeSerializer):
-            class NestedGroupSerializer(GroupSerializer):
-                class Meta:
-                    model = Group
-                    fields = ["id"]
-
-            class Meta:
-                model = RoundScope
-                fields = ["group", "vaccine"]
-
-            group = NestedGroupSerializer()
-
-        class Meta:
-            model = Round
-            fields = [
-                "id",
-                "number",
-                "started_at",
-                "ended_at",
-                "scopes",
-                "vaccine_names",
-                "target_population",
-                "is_planned",
-            ]
-
-        def to_representation(self, instance):
-            # Skip rounds on hold
-            if instance.on_hold:
-                return None
-            return super().to_representation(instance)
-
-    class NestedScopeSerializer(CampaignScopeSerializer):
-        class NestedGroupSerializer(GroupSerializer):
-            class Meta:
-                model = Group
-                fields = ["id"]
-
-        class Meta:
-            model = CampaignScope
-            fields = ["group", "vaccine"]
-
-        group = NestedGroupSerializer()
-
-    rounds = NestedListRoundSerializer(many=True, required=False)
-    scopes = NestedScopeSerializer(many=True, required=False)
-    campaign_types = CampaignTypeSerializer(many=True, required=False)
-    sub_activities = serializers.SerializerMethodField()
-
-    def get_sub_activities(self, campaign):
-        sub_activities = SubActivity.objects.filter(round__campaign=campaign)
-        return self.NestedSubactivitySerializer(sub_activities, many=True, context=self.context).data
-
-    def to_representation(self, instance):
-        data = super().to_representation(instance)
-        # Filter out None values from rounds (the test rounds we skipped)
-        data["rounds"] = [r for r in data["rounds"] if r is not None]
-        return data
-
-    class Meta:
-        model = Campaign
-        fields = [
-            "id",
-            "epid",
-            "obr_name",
-            "account",
-            "top_level_org_unit_name",
-            "top_level_org_unit_id",
-            "rounds",
-            "sub_activities",
-            "is_preventive",
-            "general_status",
-            "grouped_campaigns",
-            "separate_scopes_per_round",
-            "scopes",
-            # displayed in RoundPopper
-            # To deprecate in front-end code as it doesn't have subactivities
-            "vaccines",
-            "single_vaccines",
-            "campaign_types",
-            "description",
-            "is_test",
-            "on_hold",
-            "is_planned",
-        ]
-        read_only_fields = fields
-
-
-def preparedness_from_url(spreadsheet_url, force_refresh=False):
-    try:
-        if force_refresh:
-            ssi = SpreadSheetImport.create_for_url(spreadsheet_url)
-        else:
-            ssi = SpreadSheetImport.last_for_url(spreadsheet_url)
-        if not ssi:
-            return {}
-
-        cs = ssi.cached_spreadsheet
-        r = {}
-        preparedness_data = get_preparedness(cs)
-        r.update(preparedness_data)
-        r["title"] = cs.title
-        r["created_at"] = ssi.created_at
-        r.update(get_preparedness_score(preparedness_data))
-        r.update(preparedness_summary(preparedness_data))
-        return r
-    except InvalidFormatError as e:
-        raise serializers.ValidationError(e.args[0])
-    except APIError as e:
-        raise serializers.ValidationError(e.args[0].get("message"))
-
-
-class PreparednessPreviewSerializer(serializers.Serializer):
-    google_sheet_url = serializers.URLField()
-
-    def validate(self, attrs):
-        spreadsheet_url = attrs.get("google_sheet_url")
-        return preparedness_from_url(spreadsheet_url, force_refresh=True)
-
-    def to_representation(self, instance):
-        return instance
-
-
-def get_current_preparedness(campaign, roundNumber):
-    try:
-        round = campaign.rounds.get(number=roundNumber)
-    except Round.DoesNotExist:
-        return {"details": f"No round {roundNumber} on this campaign"}
-
-    if not round.preparedness_spreadsheet_url:
-        return {}
-    spreadsheet_url = round.preparedness_spreadsheet_url
-    return preparedness_from_url(spreadsheet_url)
-
-
-class CampaignPreparednessSpreadsheetSerializer(serializers.Serializer):
-    """Serializer used to CREATE Preparedness spreadsheet from template"""
-
-    campaign = serializers.PrimaryKeyRelatedField(queryset=Campaign.objects.all(), write_only=True)
-    round_number = serializers.IntegerField(required=False)
-    url = serializers.URLField(read_only=True)
-
-    def create(self, validated_data):
-        campaign = validated_data.get("campaign")
-        round_number = validated_data.get("round_number")
-
-        spreadsheet = generate_spreadsheet_for_campaign(campaign, round_number)
-
-        return {"url": spreadsheet.url}
 
 
 class CampaignViewSet(ModelViewSet):
