+++ plugins/polio/api/preparedness/serializers.py
@@ -1,0 +1,37 @@
+from rest_framework import serializers
+
+from plugins.polio.api.preparedness.utils import preparedness_from_url
+from plugins.polio.models import (
+    Campaign,
+)
+from plugins.polio.preparedness.spreadsheet_manager import (
+    Campaign,
+    generate_spreadsheet_for_campaign,
+)
+
+
+class PreparednessPreviewSerializer(serializers.Serializer):
+    google_sheet_url = serializers.URLField()
+
+    def validate(self, attrs):
+        spreadsheet_url = attrs.get("google_sheet_url")
+        return preparedness_from_url(spreadsheet_url, force_refresh=True)
+
+    def to_representation(self, instance):
+        return instance
+
+
+class CampaignPreparednessSpreadsheetSerializer(serializers.Serializer):
+    """Serializer used to CREATE Preparedness spreadsheet from template"""
+
+    campaign = serializers.PrimaryKeyRelatedField(queryset=Campaign.objects.all(), write_only=True)
+    round_number = serializers.IntegerField(required=False)
+    url = serializers.URLField(read_only=True)
+
+    def create(self, validated_data):
+        campaign = validated_data.get("campaign")
+        round_number = validated_data.get("round_number")
+
+        spreadsheet = generate_spreadsheet_for_campaign(campaign, round_number)
+
+        return {"url": spreadsheet.url}
