+++ plugins/polio/api/preparedness/utils.py
@@ -1,0 +1,46 @@
+from gspread.exceptions import APIError  # type: ignore
+from rest_framework import serializers
+
+from plugins.polio.models import (
+    Round,
+    SpreadSheetImport,
+)
+from plugins.polio.preparedness.calculator import get_preparedness_score
+from plugins.polio.preparedness.parser import InvalidFormatError, get_preparedness
+from plugins.polio.preparedness.summary import preparedness_summary
+
+
+def preparedness_from_url(spreadsheet_url, force_refresh=False):
+    try:
+        if force_refresh:
+            ssi = SpreadSheetImport.create_for_url(spreadsheet_url)
+        else:
+            ssi = SpreadSheetImport.last_for_url(spreadsheet_url)
+        if not ssi:
+            return {}
+
+        cs = ssi.cached_spreadsheet
+        r = {}
+        preparedness_data = get_preparedness(cs)
+        r.update(preparedness_data)
+        r["title"] = cs.title
+        r["created_at"] = ssi.created_at
+        r.update(get_preparedness_score(preparedness_data))
+        r.update(preparedness_summary(preparedness_data))
+        return r
+    except InvalidFormatError as e:
+        raise serializers.ValidationError(e.args[0])
+    except APIError as e:
+        raise serializers.ValidationError(e.args[0].get("message"))
+
+
+def get_current_preparedness(campaign, roundNumber):
+    try:
+        round = campaign.rounds.get(number=roundNumber)
+    except Round.DoesNotExist:
+        return {"details": f"No round {roundNumber} on this campaign"}
+
+    if not round.preparedness_spreadsheet_url:
+        return {}
+    spreadsheet_url = round.preparedness_spreadsheet_url
+    return preparedness_from_url(spreadsheet_url)
